+++ test/test_commands.rb
@@ -15,8 +15,8 @@ class TestCommands < Minitest::Test
 
   def test_create_login_key
     @client.response = { "error" => "0" }
-    assert @client.create_login_key("test", "secret")
 
+    assert @client.create_login_key("test", "secret")
     assert_equal :post, @client.method
     assert_equal "/CMD_API_LOGIN_KEYS", @client.url
     assert_equal "test", @client.params["keyname"]
@@ -29,6 +29,15 @@ class TestCommands < Minitest::Test
     assert_equal :post, @client.method
     assert_equal "/CMD_API_VERIFY_PASSWORD", @client.url
     assert_equal "admin", @client.params["user"]
+    assert_equal "secret", @client.params["passwd"]
+  end
+
+  def test_email_auth
+    @client.response = { "error" => "0" }
+
+    assert @client.email_auth("user@domain.com", "secret")
+    assert_equal "/CMD_API_EMAIL_AUTH", @client.url
+    assert_equal "user@domain.com", @client.params["email"]
     assert_equal "secret", @client.params["passwd"]
   end
 end