+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # master
 - Verify Password command
+- Email Auth command
 - Better tests
 
 # 0.1.0 (2017-04-15)